child non assenting enrollment e2e test
