+++ playwright-e2e/data/fake-user.json
@@ -61,7 +61,7 @@
       "DD": "20",
       "YYYY": "2016"
     },
-    "age": "06",
+    "age": "04",
     "country": {
       "name": "UNITED STATES",
       "abbreviation": "US"
